New design
I made a few changes in the original design. It's now for the original
part which should be socketed and by doing so you can easily replace it
if any fatal failure occurs. Also it uses less jumperlinks (only by one
but it's worth mentioning). This version is currently under build and
not tested yet, but it should be identitical to the previous one in the
means of function.